# PERF: Avoid duplicate mean calculations in skew and kurtosis