#### Reference issue
Follow up from gh-13254

#### What does this implement/fix?
`skew` and `kurtosis` currently recompute the same mean 3 times each: once for the scale test added in gh-13254 and two for each call to `moment`. This adds an optional `mean` argument to the internal `_moment` helper function, so we can reuse the same mean calculation between calls.

#### Additional information
Benchmark results show a 1.4-2x improvement across the board.
```
       before           after         ratio
     [838d5726]       [7d7e4210]
     <master>         <moment-optimization>
-        193±20μs          141±6μs     0.73  stats.BenchSkewKurtosis.time_kurtosis(3, 1000, False)
-        188±20μs         134±10μs     0.71  stats.BenchSkewKurtosis.time_skew(8, 10000, True)
-        189±20μs          130±5μs     0.69  stats.BenchSkewKurtosis.time_kurtosis(8, 1000, False)
-         184±7μs          123±7μs     0.67  stats.BenchSkewKurtosis.time_kurtosis(3, 100, False)
-         257±8μs          170±9μs     0.66  stats.BenchSkewKurtosis.time_kurtosis(3, 10000, False)
-         192±9μs          126±4μs     0.66  stats.BenchSkewKurtosis.time_kurtosis(3, 10000, True)
-        258±10μs          169±7μs     0.66  stats.BenchSkewKurtosis.time_kurtosis(2, 10000, False)
-        264±30μs          170±6μs     0.64  stats.BenchSkewKurtosis.time_kurtosis(1, 10000, False)
-        196±20μs          125±5μs     0.64  stats.BenchSkewKurtosis.time_kurtosis(2, 10000, True)
-        195±10μs          124±3μs     0.64  stats.BenchSkewKurtosis.time_skew(3, 10000, True)
-        221±30μs          140±8μs     0.64  stats.BenchSkewKurtosis.time_skew(1, 1000, False)
-        188±10μs          119±9μs     0.63  stats.BenchSkewKurtosis.time_kurtosis(8, 10000, True)
-        280±40μs          176±3μs     0.63  stats.BenchSkewKurtosis.time_skew(1, 10000, False)
-        205±20μs          128±6μs     0.62  stats.BenchSkewKurtosis.time_kurtosis(1, 1000, False)
-        195±10μs          121±8μs     0.62  stats.BenchSkewKurtosis.time_kurtosis(1, 10000, True)
-         153±6μs        93.6±10μs     0.61  stats.BenchSkewKurtosis.time_skew(2, 1000, True)
-         127±7μs         77.2±4μs     0.61  stats.BenchSkewKurtosis.time_kurtosis(1, 100, True)
-        145±20μs         87.1±5μs     0.60  stats.BenchSkewKurtosis.time_kurtosis(2, 1000, True)
-        136±10μs         81.9±9μs     0.60  stats.BenchSkewKurtosis.time_kurtosis(2, 100, True)
-         142±7μs         84.2±4μs     0.59  stats.BenchSkewKurtosis.time_kurtosis(3, 1000, True)
-        151±10μs         89.0±6μs     0.59  stats.BenchSkewKurtosis.time_kurtosis(8, 1000, True)
-        128±20μs         75.0±4μs     0.59  stats.BenchSkewKurtosis.time_kurtosis(3, 100, True)
-        130±10μs         76.2±4μs     0.59  stats.BenchSkewKurtosis.time_kurtosis(8, 100, True)
-        184±40μs         90.5±7μs     0.49  stats.BenchSkewKurtosis.time_kurtosis(1, 1000, True)

SOME BENCHMARKS HAVE CHANGED SIGNIFICANTLY.
PERFORMANCE INCREASED.

```